refactor(archive): own tail overlay metadata
