+++ polylogue/archive/conversation/runtime.py
@@ -8,11 +8,11 @@ from datetime import datetime
 from typing import TYPE_CHECKING, Self, cast
 
 from polylogue.archive.conversation.branch_type import BranchType
+from polylogue.archive.conversation.tail_overlay import TailOverlayInfo, tail_overlay_from_provider_meta
 from polylogue.archive.message.messages import MessageCollection
 from polylogue.archive.message.models import DialoguePair, Message
 from polylogue.archive.message.roles import normalize_message_roles
 from polylogue.lib.roles import Role
-from polylogue.lib.tail_overlay import TailOverlayInfo, tail_overlay_from_provider_meta
 from polylogue.types import ConversationId
 
 if TYPE_CHECKING:
