+++ polylogue/archive/conversation/summary_runtime.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from datetime import datetime
 
 from polylogue.archive.conversation.branch_type import BranchType
-from polylogue.lib.tail_overlay import TailOverlayInfo, tail_overlay_from_provider_meta
+from polylogue.archive.conversation.tail_overlay import TailOverlayInfo, tail_overlay_from_provider_meta
 from polylogue.types import ConversationId
