+++ polylogue/surfaces/payloads.py
@@ -19,9 +19,9 @@ if TYPE_CHECKING:
     from collections.abc import Container
 
     from polylogue.archive.conversation.neighbor_candidates import ConversationNeighborCandidate, NeighborReason
+    from polylogue.archive.conversation.tail_overlay import TailOverlayInfo
     from polylogue.archive.query.search_hits import ConversationSearchHit
     from polylogue.lib.models import Conversation, ConversationSummary, Message
-    from polylogue.lib.tail_overlay import TailOverlayInfo
 
 
 def serialize_surface_payload(payload: BaseModel, *, exclude_none: bool = False) -> str:
